Add a tool to generate firmware extracting scripts for xc3028
From: Mauro Carvalho Chehab <mchehab@infradead.org>

This is the tool used to generate extract_xc3028.pl.

Once you have a firmware file, generated from other methods, like usb dumps,
this tool allows to seek the firmware inside a windows driver, creating an
extrating script.

There are two methods to seek for a firmware:
	1) a complete firmware seek;
	2) seek for firmware hunks. For each firmware segment, it writes the
segment size, and seeks for the segment inside the firmware.

The result of method 2) is not optimized.

In the case of hvr12x0 driver, the base firmwares didn't hit method (1), so,
the extracting tool broke it into several small segments. This is not
optimized, but works.

Signed-off-by: Mauro Carvalho Chehab <mchehab@infradead.org>