v4l2-ctl: fix bug in --set-standard
From: Hans Verkuil <hverkuil@xs4all.nl>

--set-standard allows you to set a standard index corresponding to
what ENUMSTD returns. But this clashes when you specify a standard whose
ID is very small, like PAL-BG. Then the utility thinks you specified
a standard index and will select the wrong standard.

Fixed by using a flag to mark that the standard actually is a standard
and not a standard index.

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>