v4l2: debugging API changed to match against driver name instead of ID.
From: Hans Verkuil <hverkuil@xs4all.nl>

Since the i2c driver ID will be removed in the near future we have to
modify the v4l2 debugging API to use the driver name instead of driver ID.

Note that this API is not used in applications other than v4l2-dbg.cpp
as it is for debugging and testing only.

Should anyone use the old VIDIOC_G_CHIP_IDENT, then this will be logged
with a warning that it is deprecated and will be removed in 2.6.30.

Priority: normal

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>