v4l2grab: Add an userspace application for webcam testing, using libv4l2
From: Mauro Carvalho Chehab <mchehab@redhat.com>

During FISL 10 (Forum Internacional de Software Livre) and the first FudCon Latam
(Fedora Users and Developers Conference), while preparing the V4L panel, I and Douglas
noticed the need of a very simple application to be an example for developers to
test their drivers and to write newer v4l2 applications.

So, we come to the idea of creating a v4l2grab application that takes pictures from the
webcam and saves them with the ppm format, using libv4l to work with all webcam models.

This application uses the absolute minimum of V4L2 calls to take 640x480 pictures, and was tested
with uvcvideo and vivi drivers.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>