v4l2-apps: add qv4l2 for Qt4
From: Hans Verkuil <hverkuil@xs4all.nl>

This Qt4 version of qv4l2 is much improved compared to the Qt3 version.
It adds support for the libv4l wrappers and it can capture as well.

There are still many improvements that can be made, but it is time to
get this version merged so people can start playing with it.

Priority: normal

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>