v4l2-ctl: fix regression in ability to set/get private controls
In hg revision 12546, a regression was introduced which resulted in the
ability to get/set private controls.  The change resulted in all attempts
to set private controls going through the extended controls interface, and
the extended controls interface explicitly denies ability to use private
control CIDs (it's enforced in the check_ext_ctl function in v4l2-ioctl.c.

Fix the code such that it goes back to using the older g_ctrl/s_ctrl if the
control ID is a private control.

Priority: high

Signed-off-by: Devin Heitmueller <dheitmueller@kernellabs.com>
Acked-by: Hans Verkuil <hverkuil@xs4all.nl>