libv4l2: Ignore convert errors in the first few frames of a stream
Rather then reporting an error to the caller when converting one of
the first frames of the stream fails, get another frame and try again.

Some cams (ie UVC JPEG cams) produce a bad frame when the stream starts
(with for example a bad or missing jpeg header). We used to report this
as an error to the caller which causes programs which quit on the
first error (such as camorama) to terminate.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>