keytable: Parse also the IR_TABLE macro
The IR_TABLE macro provides two useful information: the type
of the IR table, and the table name, as reported by kernel uevent,
at node /sys/class/irrcv/irrcv0/uevent. For example:
	NAME="rc5_hauppauge_new"

This information can be useful to allow userspace to know that
a table needs to be replaced by another.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>