qv4l2: move qv4l2-qt3 to contrib, rename qv4l2-qt4 directory to qv4l2.
qv4l2-qt3 is effectively unmaintained. Move it to contrib and rename the
executable to qv4l2-qt3. If someone needs a Qt3 tool, then they can use this.

The qv4l2-qt4 directory can now be renamed to qv4l2 and the Makefile is
modified to only build if qt4 is installed.

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>