qv4l2: fix memory corruption and a wrong close sequence
When allocating memory for a string control the wrong size was used.

When closing the video device (e.g. during exit) the signal mapper was
deleted after the dialog was destroyed. This caused lostFocus events to arrive
in the signal mapper's handler which then tried to access the already closed
video device. And that caused a confusing error message to appear.

This is now fixed by deleting the signal mapper first.

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>