ir-keytable: Put all driver values into a struct
Now, instead of a series of static vars, all device information
is stored at rc_dev struct. With this change, get_attribs will
now fill the struct with all information that were retrieved from
the device sysfs nodes.

With this change, it is now easier to get additional info from the
RC sysfs nodes, and to provide a set_attribs function to change
data, like the supported protocols.

As a bonus, the list mode now provides more info about the RC
devices.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>