libv4l: add detection for short frames
Sometimes cam generate short frames (esp. at the beginning of the stream),
detect this and retry a number of times. If retrying fails, pass allong
the short frame as it is better then nothing.

This patch also cleans up the JPEG decoding error handling code to use
the same mechanism.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>