qv4l2: Add a desktop description for qv4l2 and new icons.
As we intend to have qv4l2 application installed on distros, we
need a .desktop description for it, plus icons of different sizes.

Created a svg icon close to the existing one, and added bitmaps
for it with different resolutions typically found on modern X
menus (16x16, 24x24, 32x32 and 64x64 resolutions). All of them
derivated from the svg file, and generated using inkscape.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>