Add an initial parser for usbmon captured via wireshark
Usbmon allows capturing usb driver with tcpdump or wireshark.
However, data may need to be parsed, in order to process them
by some script.

The easiest way to parse is to use wireshark, exporting the output
dump with packet details, all expanded, and use a parser to convert
it into an easier format.

This parser is able to decode wireshark parsed files. For now,
it will only print the parsed data.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>