Parse usbmon tcpdump logs directly
Instead of relying on wireshark parser, just read the tcpdump
logs directly, and use it to produce a parseable output.

I started to use my Beagleboard as an USB sniffer using the
code at:
	http://beagleboard-usbsniffer.blogspot.com/

However, as I want to use the parser scripts for USB that I wrote,
I decided to write a parser that outputs data on a similar format.

On a first trial, I tried to parse the wireshark dump, but this
didn't work fine. So, I rewrote my parser to get the tcpdump
directly, using the cpan Net::TcpDumpLog.

There are some adjustments probably needed, but this version
already produces something usefull.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>