parse_em28xx.pl: Parse also i2c device scans
While comparing the parsed messages between the saved register data
of em28xx and netmon parser, noticed that the scan messages were not
parsed. Also, it seems that the data from the driver is wrong, as it
is saying that some data were returned by the probed devices.

Comparing both logs is actually a very good test, as it shows that the
netfilter parser is working properly, as the only differences are on
the I2C scan messages.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>