parse_tcpdump_log.pl: Remove TcpDumpLog.pm dependency
TcpDumpLog.pm does a good job parsing tcpdump logs. However, it
works by reading everything into the memory, before allowing to
parse the data. This is not very good when dealing with very
long logs that could be produced by an USB 2.0 data stream device.

So, instead, let's copy the TcpDumpLog parser into the code, and
change it to work line by line. One additional advantage of using
such approach is that we could change the parser to also start
the usbmon monitoring, allowing to do a real time parsing.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>