parse_tcpdump_log.pl: add support for collecting data directly
Instead of needing to use tcpdump, just use Net:Pcap and retrieve the
data directly via libpcap. This allows real-time parsing of the logs,
and may be very useful while debugging certain drivers.

The usual way via tcpdump is not touched.

A patch is needed in order to optionally use Net::Pcap, as this is only
required on pcap mode.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>