libmedia_dev: Create a new library to handle media devices
The libmedia_dev library contains API's meant to be used by media
applications that needs to share some common code.

For now, it contains the routines that enumerates the media devices
and provides the associated audio devices that are needed in order
to allow playing audio for a given V4L source.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>