get_media_devices: Make internal structs opaque
The internal structs used to retrieve sysfs information were
designed to fulfill internal design requirements in a way that
the data is easily stored inside it. They're not meant to
be used outside the library. Other methods will export the
retrieved information in a way that they'll be useful for
the applications.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>