qv4l2: don't change the format just because we use libv4l2.
So some reason qv4l2 switched to the RGB24 format if the video node was
opened through libv4l2. This is a completely unexpected side-effect that
should not happen, especially since qv4l2 is usually used as a test
utility.

This may mean that a frame is converted twice: once inside libv4l2, then
to get it into a format suitable for the screen. But that's OK. After all,
that's what you asked for.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>