qv4l2: add a menu option to turn off the Capture window.
Converting frames into a format suitable for displaying can take time.
If you just want to test the framerate of the capture device then it is
useful to be able to just capture without doing any format conversions
and displaying the frames.

Add a new option in the menu to do just this.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>