Allow to configure per library subdirectories for helpers.
In Debian (and Ubuntu) libv4lconvert is moved into it's own
library package due to the SONAME bump. The policy requires
that subdirectories beneath /usr/lib needs to be named after
the package name to allow older and newer package versions
installed side by side.

This patch adds the possibility to specify the subdirectories
separately for libv4lconvert, libv4l1 and libv4l2.

Signed-off-by: Gregor Jasny <gjasny@googlemail.com>