Improved JL2005BCD patch for libv4lcnvert
Hans,

The patch below supercedes the previous one I sent.

Two particular items of interest:

1. The Bayer tiling question seems to be resolved.

2. White balance is enabled and seems to help a lot. I am not certain if
one ought to enable a gamma setting as well, but things look pretty decent
without it, seems to me.

Other remarks:

It is pretty clear that these cameras have an internal gain setting which
relies upon internal feedback for self-correction. It works pretty well
most of the time, but sometimes it overcorrects and thereby goes a little
bit crazy.

I have tried it out with a bunch of cameras now. I did not check
individually which one is which, but I have a fairly large set of them
which includes the B and C and D variants, and some of the fancy ones have
the LCD previewer on the back but definitely do not send preview
thumbnails in webcam mode. So all of them seem to work. I would dare to
presume that your camera will work, too, whenever you get around to
checking it.

Patch below created by git diff from my recently downloaded git tree, as
was the last one. Patch is

Signed-off-by Theodore Kilgore <kilgota@auburn.edu>