utils/dvb: Start a dvbv5 scan application
This application is based on dvbv5-zap application, plus a reading
at the DVB specs. It might be easier to start from some other DVB
scan tool, but reading from specs will help to be sure that it will
follow the latest standards. It also makes easier to develop a
DVB-API independent version of the library, and then write some code
to convert from its internal representation into the DVBv5
specifics.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>