utils/dvb: use standard "channels.conf" file
Instead of using ~/.tzap/channels.conf file, uses a standard
channels file, without the PID's. Also, pass it as a parameter,
just like dvb-apps "scan" tool.

After this patch, the tool is working fine for ISDB-T. It
will likely work fine also for DVB-T.

Additional changes are required for DVB-C, in order to proccess
the NIT/SDT tables there, and for DVB-S, due to satellite
specifics.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>