libv4l2: Never be lazy on s_fmt
Before this patch libv4l2 tries to avoid sending an "unnecessary" s_fmt to
the device. So if an app does a s_fmt with parameters which exactly match
what is already in place libv4l2 used to not send the s_fmt to the device.

This means however that if the device was set to a specific framerate
with a s_parm call, that it will stay at this framerate. Where as
normally a s_fmt resets the device to its default framerate.

This patch ensures that app which do a s_fmt always get the default
framerate by always forwarding the s_fmt to the device.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>