qv4l2: Do a set_interval after s_fmt
s_fmt resets a device to its default interval, so if a specific
interval has been selected we should so a set_interval after a s_fmt to
make sure we actually get the specified interval.

Also don't enable the drop down if enum_interval returns a non discrete
interval, since we're not populating the drop down then.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>