dvb/descriptors: improve ISDB-T verbosity
On ISDB-T, most of the network parameters are decoded via the TMCC
carrier, modulated on all segments using DBPSK. It is up to the
ISDB-T demod to parse the TMCC registers and pass their contents via
get_frontend().

Yet, the NIT table provides a few data:
	ISDB-T mode (1, 2 or 3);
	Guard interval.

The ISDB-T mode is directly related to the number of OFDM carriers,
and the IFFT size:
	Mode 1 corresponds to an IFFT size of 2K;
	Mode 2 corresponds to an IFFT size of 4K;
	Mode 3 corresponds to an IFFT size of 8K.

Provide a better output of those data, if verbose mode is enabled.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>