dvb-file: Allow writing a zap/channel file for a DVBv5-only standard
ISDB-T is not really supported by a DVB-T format. Yet, DVB-T tools
are able to work with with an ISDB-T frontend in compat mode.

So, while the conversion to a legacy format is not a rocket science,
it is interesting to allow writing/converting a DVB-T file for an
ISDB-T scan file.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>