libv4lconvert: fix setting/testing of supported_src_formats
data->supported_src_formats is a uint64_t.

This code:

data->supported_src_formats & (1 << i)

does not work since (1 << i) is a 32-bit value. It should be (1ULL << i) instead.

Interesting fact: on a 64-bit intel box (1 << 33) does a rotate, not a shift.
So (1 << 33) == 2.

On a 32-bit powerpc platform it is a proper shift, so (1 << 33) == 0.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>