libv4l2: Ensure we always set buf->length when converting
Some apps blindly use buf->length when calling munmap, even if a previous
call (ie dqbuf) on the buffer failed. We used to not set buf->length to
our conversion buffer length when the actual ioctl call to the device, or
the format conversion failed.

When a dqbuf succeeded and the conversion failed this which would cause
us to leave buf->length set to the real buffer length rather then the
conversion buffer length, if the app then wrongly uses buf->length (*)
for a munmap we would not recognize this as munmap of the conversion
buffer and pass it through to the real munmap unmapping a part of
our conversion buf without being aware of this!

*) After from the point of the app a failed dqbuf, so it should not trust the
contents of buf at all.

This patch makes things work even for buggy apps, by always setting
buf->length when converting even if things fail.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>

Conflicts:
	lib/libv4l2/libv4l2.c