v4l2-compliance: Not Supported -> OK (Not Supported)
The 'Not Supported' result was confusing: this is not an error, it
just says that the tested API is not implemented in the driver, and that's
OK since that API is optional.

Renaming this to 'OK (Not Supported)' should clarify this result.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
(cherry picked from commit 98c150227b8e27702ce397351cfb1ccf33ab5123)

Signed-off-by: Gregor Jasny <gjasny@googlemail.com>