libv4l: Symlink wrapper libraries to libdir
The curent approach to install the wrapper libraries into a @prefix@/lib/
subdirectory has the downside that the user has to specify the full path
for preloading. Finding out the correct path is sometimes not that easy,
especially if crossarch or multiarch are involved.

With that commit a symlink is created and the loader should pick the
suitable library itself.

From now on it should be enough to run "LD_PRELOAD=v4l2convert.so vlc"
instead of specifying the full path.

I chose symlinks to the private dir and not the other way round to make
the wrappers co-installable across multiple SONAME generations.

Signed-off-by: Gregor Jasny <gjasny@googlemail.com>