v4l2-compliance: mark non-NULL priv as an error.
Although the spec allows the priv field of v4l2_pix_format to be non-NULL,
in practice this is almost certainly an error. There should be no reason for
drivers to actually use this field.

We probably should look at changing the spec, possibly freeing up at least one
field for other use.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>