v4l2-ctl: add --stream-loop for streaming to an output.
When streaming from a file to an output it can be useful to loop the file
when the end is reached. Add the --stream-off option to do that.

Also improve the code to stop when a partial frame is read instead of pointlessly
trying to read data from the filehandle even if there isn't any.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>