dvb-fe: Add a flag to indicate if DVBv5 stats is in use
Instead of trying every time to see if DVBv5 stats are available,
if version is equal or upper than 5.10, a call to FE_GET_PROPERTY
will fail only if a more permanent error happened (EFAULT, ENOMEM,
and similar errors).

A successful call to FE_GET_PROPERTY that returns all stat props
with len=0 means that the frontend doesn't implement any DVBv5
stats. So, if it fails once, it can be assumed that there's no
DVBv5 stats implemented for that frontend.

As a bonus, the logic now knows exactly if a measure came from
DVBv3 or from DVBv5 API.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>