v4l2-* utils: allow -d11 to refer to /dev/video11
Lately I have been working with devices with lots of video nodes. Various
v4l2 utilities allow -dX to refer to /dev/videoX, but 'X' was limited to
a single digit. Extend this to support up to three digits. So -d11 would
refer to /dev/video11 and -d128 to /dev/video128 (that really should be
enough!)

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>