v4l2-compliance: improve tests when TRY/S_FMT returns EINVAL.
If a wrong pixelformat is specified, then TRY/S_FMT can return EINVAL.
However, part of the test sequence was skipped in that case, which led
to missed errors.

Now get a valid pixelformat and put that in and try again. That way all
other tests are executed as well.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>