v4l2-ctl: add eos and decoder-stop support.
When dealing with compressed streams it is useful if the capture streaming
code would check for V4L2_EVENT_EOS and if the output streaming side would
be able to use a STOP DECODER command before calling STREAMOFF to give the
decoder the time to flush any pending data.

This patch adds support for that.

Thanks to Tzu-Jung Lee for testing this.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>