v4l2-ctl: poll/non-poll behavior fixes
The 'v4l2-ctl: add eos and decoder-stop support.' commit broke non-poll
streaming. This is now fixed.

In addition, the eos handling didn't work either in the non-poll case and
we still want to get hold of the eos event. So poll the exception state
even in the non-poll case.

Finally, instead of using + and - to denote capture and output frames use
'<' and '>' respectively. This matches the semantics of < and > in sh.

And I also find it easier to read.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>