v4l2-compliance: refactor code to create invalid formats.
In order to test the format ioctls we fill v4l2_format with invalid
values. But for the overlay buffer types the clips and bitmap fields
must be NULL. This was done in one place, but not in the others.

Make a function that does this and call it in all places where it
is needed.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>