qv4l2: always update controls by default
The 'Update on change' checkbox was by default only checked for the user controls
tab, not for the others. This is however very confusing and in general you always
want to update a control immediately once it is changed, unless you explicitly
uncheck the checkbox.

The original idea was that for e.g. the MPEG controls you want to set them all up
just the way you want it, and then press Update to update all the changes using
extended controls. This is particularly useful for controls that depend on one
another.

In practice however, and certanly with lots of the new controls that appeared,
you just want to have an immediate effect when changing a control.

Having 'Update on change' disabled just leads to confused users, which frankly
includes myself :-)

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>