libdvbv5: remove the remaining bits from the NIT parsers
The previous NIT parser used to fill the entries data based on
the information detected on each transponder/channel.

Well, if the frontend was able to lock on the transponder, that
means that the parameters there at the entry are already correct
for the current channel, and that there's no need to use the NIT
table.

For new transponders frequencies, those data should be applied,
but this is done already inside dvb_add_scaned_transponders().

So, remove this code. As this is the last part using the old
descriptors logic, remove the struct definitions as well.

Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>