libdvbv5: better implement the NIT parser
The code that were implementing the NIT parser were becoming too
complex, and had some common duplicated code.

Re-implement it, creating a helper routine to parse any NIT data,
and re-using the same routine to allow updating both the current
transponder and the detected transponders on a cleaner and saner
way.

Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>