libdvbv5: Simplify PAT handling
Coverity thinks that there's a resource leak by having a pgm
not being properly stored.

While I don't agree with Coverity, the code could be simplified,
and use just one pointer instead of two, in order to fill the
programs linked list.

That makes it easier to be analyzed by both automatic tools and
by manual inspection.

So, simplify it.

Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>