libdvbv5: Only discard the services not found
If the PAT table describes a service ID, but it doesn't exist, just
discard it, and not all services after that.

This helps to be a little more resillient to broken PAT tables.

Caught with a test signal that was captured by a friend in Rio. The
tables describe a SD stream that it is not present on the TS.

$ dvbv5-scan ~mchehab/isdb-teste -F -G -T2
INFO     Scanning frequency #1 479142857
Lock   (0x1f) Signal= 78.83% C/N= 0.38% UCB= 252 postBER= 0
INFO     Service Globo HD, provider (null): digital television
INFO     Service Globo SD, provider (null): digital television
WARNING  Service ID 59201 not found on PMT!
INFO     Service Globo 1Seg, provider (null): user defined

Not sure if this is due to some capture error or not, but it is better
to be able to capture the 1Seg stream in any case, and just discard the
broken Service ID.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>