dvb-scan: don't handle NIT table by default on ATSC
With ATSC, the NIT table data is not useful at scan, as the only
non-automatically-detected parameter is the modulation. Well, the
modulation is mandatory anyway.

So, the scanning tool don't need the NIT table at all.

As NIT table is not always present (it isn't in KR), that makes
the scan procedure faster.

We could eventually disable it for other delivery systems, but
that would require an extra parameter on two functions, and this
would break userspace API. So, let's apply this patch instead.

Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>