dvb-fe: Add support for setting the LNA before tuning
Sometimes, we need to control the Low Noise Amplifier (LNA)
at the antena, in order to get a signal. This is dependent on
the way the tuner is connected to the antena, and if the antena
has support for it.

Add a logic at the code to set it. By default, LNA will be in auto
mode, with is the DVB core's default.

Enabling or disabling it is just a matter of changing parms->lna.

Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>