v4l2-compliance: add EXPBUF and DMABUF streaming tests.
At the moment only single planar VIDEO_CAPTURE buffer types are supported.

To test DMABUF streaming you need a /dev/video device that supports
VIDIOC_EXPBUF. In addition, that device needs to be set to a format that
creates large enough buffers to support the device under test.

You can specify which device to use by passing the --expbuf-device option.

Devices that support EXPBUF are generally found in embedded drivers,
however work is in progress to support it in the saa7134 driver and that
is what I have been using to test this.

Hardware using the saa7134 is easy to find, so this is a reasonable
option.

It would be nice if we could also use GPU drivers for this, but someone
with more knowledge than myself needs to help out for that.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>