v4l-utils: fix optional args for short opts
I always wondered why I couldn't do e.g. -s10 where 10 is the optional
argument for the short option -s. It worked fine with --streaming=10.
It turns out that in the optstring passed to getopt_long you have to
specify such an option as 's::' and I didn't so that.

This is now fixed.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>