v4l2-ctl: improve the overlay support
It is now possible to specify clip lists and a bitmap.

With the --find-fb option it can find which fb device is associated with
the overlay.

And --set-fbuf can specify which fb device to open to fill in the framebuffer
information.

Dropped the old output overlay options: each video node can only be a video
overlay or an output video overlay, not both at the same time. So there is
no need to support both options.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>