qv4l2: fix incorrect call to s_fmt when disabling width/height
When the frame width and height fields are disabled because we start
streaming, the editingFinished event is triggered as well, which
causes an attempt to change the format. This causes EBUSY warnings
because it really isn't allowed to do that anymore. Check if the
field is really enabled before we change the format, thus squashing
the warnings.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>