libv4lconvert: remove broken ALTERNATE handling
The V4L2 specification used to say that if field == V4L2_FIELD_ALTERNATE, the
height would have to be divided by two. This is incorrect, the height is that of
a single field. This has been corrected in the spec, now this code in libv4lconvert
needs to be removed as well.

Tested with both bttv and saa7146, the only two drivers supporting FIELD_ALTERNATE
today.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>